Delete unused TaskApiService.cs, Move handle expiry to ResultsService.cs & update Aggregate method
