+++ app/Hutch.Relay/Services/Hosted/TaskCompletionHostedService.cs
@@ -17,28 +17,11 @@ public class TaskCompletionHostedService(
     while (!stoppingToken.IsCancellationRequested)
     {
       using var scope = serviceScopeFactory.CreateScope();
-      var relayTaskService = scope.ServiceProvider.GetRequiredService<IRelayTaskService>();
       var taskResultsService = scope.ServiceProvider.GetRequiredService<ResultsService>();
-
-      logger.LogInformation("Checking for incomplete Tasks...");
-      var incompleteTasks = await relayTaskService.ListIncomplete();
-      foreach (var task in incompleteTasks)
-      {
-        logger.LogInformation("Incomplete task:{Task}", task.Id);
-        var timeInterval = DateTimeOffset.UtcNow.Subtract(task.CreatedAt);
-        if (timeInterval > TimeSpan.FromMinutes(4.5))
-        {
-          // get subtasks 
-          var subTasks = await relayTaskService.ListSubTasks(task.Id, incompleteOnly: false);
-          // aggregate subtasks
-          var aggregateCount = taskResultsService.AggregateResults(subTasks.ToList());
-          //Post back to api 
-
-          //Set task as complete
-          await relayTaskService.SetComplete(task.Id);
-        }
-      }
-
+      
+      logger.LogInformation("Finding Tasks that are expiring...");
+      
+      await taskResultsService.HandleResultsToExpire();
       await Task.Delay(TimeSpan.FromSeconds(5), stoppingToken);
     }
   }
