+++ app/Hutch.Relay/Services/ResultsService.cs
@@ -4,6 +4,7 @@ using Hutch.Rackit;
 using Hutch.Rackit.TaskApi.Contracts;
 using Hutch.Rackit.TaskApi.Models;
 using Hutch.Relay.Models;
+using Hutch.Relay.Services.Contracts;
 using Microsoft.Extensions.Options;
 
 namespace Hutch.Relay.Services;
@@ -11,8 +12,11 @@ namespace Hutch.Relay.Services;
 public class ResultsService(
   ILogger<ResultsService> logger,
   IOptions<ApiClientOptions> options,
-  ITaskApiClient upstreamTasks)
+  ITaskApiClient upstreamTasks,
+  IRelayTaskService relayTaskService)
 {
+  private ApiClientOptions options = options.Value;
+
   public async Task SubmitResults(RelayTaskModel relayTask, JobResult jobResult
   )
   {
@@ -21,11 +25,11 @@ public class ResultsService(
     int maxRetryCount = 5;
     while (retryCount < maxRetryCount)
     {
-      logger.LogInformation("Submitting Results..");
+      logger.LogInformation("Submitting Results for {Task}..", relayTask.Id);
       try
       {
         // Submit results upstream
-        await upstreamTasks.SubmitResultAsync(relayTask.Id, jobResult, options.Value);
+        await upstreamTasks.SubmitResultAsync(relayTask.Id, jobResult, options);
         logger.LogInformation("Successfully submitted results for {RelayTaskId}", relayTask.Id);
         break;
       }
@@ -45,10 +49,12 @@ public class ResultsService(
     }
   }
 
-  public int AggregateResults(List<RelaySubTaskModel> relaySubTasks)
+  public async Task<JobResult> AggregateResults(string relayTaskId)
   {
+    // Get all SubTasks for the Task
+    var subTasks = await relayTaskService.ListSubTasks(relayTaskId, incompleteOnly: false);
     int aggregateCount = 0;
-    foreach (var subTask in relaySubTasks)
+    foreach (var subTask in subTasks)
     {
       if (subTask.Result != null)
       {
@@ -57,6 +63,34 @@ public class ResultsService(
       }
     }
 
-    return aggregateCount;
+    return new JobResult()
+    {
+      Uuid = relayTaskId,
+      CollectionId = options.CollectionId ??
+                     throw new ArgumentException(nameof(options.CollectionId)),
+      Results = new QueryResult()
+      {
+        Count = aggregateCount,
+      }
+    };
+  }
+
+  public async Task HandleResultsToExpire()
+  {
+    var incompleteTasks = await relayTaskService.ListIncomplete();
+    foreach (var task in incompleteTasks)
+    {
+      logger.LogInformation("Task:{Task} is about to expire.", task.Id);
+      var timeInterval = DateTimeOffset.UtcNow.Subtract(task.CreatedAt);
+      if (timeInterval > TimeSpan.FromMinutes(4.5))
+      {
+        // Aggregate SubTask count
+        var finalResult = await AggregateResults(task.Id);
+        // Post to TaskApi 
+        await SubmitResults(task, finalResult);
+        //Set task as complete
+        await relayTaskService.SetComplete(task.Id);
+      }
+    }
   }
 }
